# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления солнечными панелями и расчета их мощности",
        "title": "Solar Panels Power Calculator API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/api",
    "paths": {
...
        "/solarpanel-requests/{id}/{solarpanelId}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет значение площади для панели в заявке",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RequestPanels"
                ],
                "summary": "Изменить площадь панели в заявке",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID солнечной панели",
                        "name": "solarpanelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новое значение площади",
                        "name": "area",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lab_internal_app_DTO.ChangeSolarPanelAreaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленная панель в заявке",
                        "schema": {
                            "$ref": "#/definitions/lab_internal_app_ds.RequestPanels"
                        }
                    },
                    "400": {
                        "description": "Некорректное значение площади",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Заявка недоступна этому пользователю",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Заявка или панель не найдены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет солнечную панель из заявки пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RequestPanels"
                ],
                "summary": "Удалить панель из заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID солнечной панели",
                        "name": "solarpanelId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Панель успешно удалена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Заявка недоступна / Панель отсутствует",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Заявка или панель не найдены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "lab_internal_app_DTO.ChangeSolarPanelAreaRequest": {
            "type": "object",
            "required": [
                "area"
            ],
            "properties": {
                "area": {
                    "type": "number"
                }
            }
        },
...
        "lab_internal_app_ds.RequestPanels": {
            "type": "object",
            "properties": {
                "area": {
                    "type": "number"
                },
                "solarPanel": {
                    "$ref": "#/definitions/lab_internal_app_ds.SolarPanel"
                },
                "solar_panel_id": {
                    "type": "integer"
                },
                "solar_panel_request_id": {
                    "type": "integer"
                }
            }
        },
        "lab_internal_app_ds.SolarPanel": {
            "type": "object",
            "properties": {
                "depth": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "efficiency": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "isDelete": {
                    "type": "boolean"
                },
                "power": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}